trading-bot/pyproject.toml
[tool.poetry]
name = "trading-bot"
version = "0.1.0"
description = "Qullamaggie momentum trading bot for US equities"
authors = ["Hanlin"]
packages = []

[tool.poetry.dependencies]
python = "^3.12"
alpaca-py = ">=0.43.0"
polygon-api-client = ">=1.14.0"
pandas = ">=2.2.0"
SQLAlchemy = ">=2.0.0"
APScheduler = ">=3.10.0"
PyYAML = ">=6.0.0"
python-telegram-bot = ">=21.0.0"
streamlit = ">=1.31.0"
requests = ">=2.31.0"
pytz = ">=2024.1"
numpy = ">=1.26.0"
alembic = ">=1.13.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
pytest-mock = ">=3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
